requirements.txt
aiogram==3.2.0
openai==1.6.1
PyYAML==6.0.1
